+++ src/lib.rs
@@ -57,15 +57,13 @@ mod content;
 mod elicitation;
 mod error;
 mod ext;
-mod maybe_undefined;
 #[cfg(feature = "unstable_nes")]
 mod nes;
-#[cfg(feature = "unstable_llm_providers")]
-mod nullable;
 mod plan;
 #[cfg(feature = "unstable_cancel_request")]
 mod protocol_level;
 mod rpc;
+mod serde_util;
 mod tool_call;
 mod version;
 
@@ -77,27 +75,20 @@ use derive_more::{Display, From};
 pub use elicitation::*;
 pub use error::*;
 pub use ext::*;
-pub use maybe_undefined::*;
 #[cfg(feature = "unstable_nes")]
 pub use nes::*;
-#[cfg(feature = "unstable_llm_providers")]
-pub use nullable::*;
 pub use plan::*;
 #[cfg(feature = "unstable_cancel_request")]
 pub use protocol_level::*;
 pub use rpc::*;
 pub use serde_json::value::RawValue;
+pub use serde_util::*;
 pub use tool_call::*;
 pub use version::*;
 
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
-use std::{
-    borrow::Cow,
-    ffi::OsStr,
-    path::{Path, PathBuf},
-    sync::Arc,
-};
+use std::sync::Arc;
 
 /// A unique identifier for a conversation session between a client and agent.
 ///
@@ -115,102 +106,5 @@ impl SessionId {
     #[must_use]
     pub fn new(id: impl Into<Arc<str>>) -> Self {
         Self(id.into())
-    }
-}
-
-/// Utility trait for builder methods for optional values.
-/// This allows the caller to either pass in the value itself without wrapping it in `Some`,
-/// or to just pass in an Option if that is what they have.
-pub trait IntoOption<T> {
-    fn into_option(self) -> Option<T>;
-}
-
-impl<T> IntoOption<T> for Option<T> {
-    fn into_option(self) -> Option<T> {
-        self
-    }
-}
-
-impl<T> IntoOption<T> for T {
-    fn into_option(self) -> Option<T> {
-        Some(self)
-    }
-}
-
-impl IntoOption<String> for &str {
-    fn into_option(self) -> Option<String> {
-        Some(self.into())
-    }
-}
-
-impl IntoOption<String> for &mut str {
-    fn into_option(self) -> Option<String> {
-        Some(self.into())
-    }
-}
-
-impl IntoOption<String> for &String {
-    fn into_option(self) -> Option<String> {
-        Some(self.into())
-    }
-}
-
-impl IntoOption<String> for Box<str> {
-    fn into_option(self) -> Option<String> {
-        Some(self.into())
-    }
-}
-
-impl IntoOption<String> for Cow<'_, str> {
-    fn into_option(self) -> Option<String> {
-        Some(self.into())
-    }
-}
-
-impl IntoOption<String> for Arc<str> {
-    fn into_option(self) -> Option<String> {
-        Some(self.to_string())
-    }
-}
-
-impl<T: ?Sized + AsRef<OsStr>> IntoOption<PathBuf> for &T {
-    fn into_option(self) -> Option<PathBuf> {
-        Some(self.into())
-    }
-}
-
-impl IntoOption<PathBuf> for Box<Path> {
-    fn into_option(self) -> Option<PathBuf> {
-        Some(self.into())
-    }
-}
-
-impl IntoOption<PathBuf> for Cow<'_, Path> {
-    fn into_option(self) -> Option<PathBuf> {
-        Some(self.into())
-    }
-}
-
-impl IntoOption<ToolCallId> for &str {
-    fn into_option(self) -> Option<ToolCallId> {
-        Some(ToolCallId::new(self))
-    }
-}
-
-impl IntoOption<serde_json::Value> for &str {
-    fn into_option(self) -> Option<serde_json::Value> {
-        Some(self.into())
-    }
-}
-
-impl IntoOption<serde_json::Value> for String {
-    fn into_option(self) -> Option<serde_json::Value> {
-        Some(self.into())
-    }
-}
-
-impl IntoOption<serde_json::Value> for Cow<'_, str> {
-    fn into_option(self) -> Option<serde_json::Value> {
-        Some(self.into())
     }
 }
